# Combine data members of IOStatus with Status