@hx235 2022-02-18
> IOStatus is now defined as a derived class of Status with
no new data, but additional methods.
> This will allow us to eventually
track the result of FileSystem calls in RocksDB with one variable
instead of two.

Just for curiosity, will the "eventual variable" used to track the result of FileSystem calls need those additional methods that are currently in IOStatus but not in Status (e.g, Get/Set functions of added protected members `retryable_`, `data_loss_` and `scope_` )? Will that "eventual variable" be the new `Status` class ("new" as in after this PR) or the new `IOStatus` class or something else?


Syntax-wise, the PR looks good to me but I will defer to more experienced reviewers for final decision.