@anand1976 2022-02-18
> 
> Will the "eventual variable" used to track the result of FileSystem calls need those additional methods that are currently in IOStatus but not in Status (e.g, Get/Set functions of added protected members `retryable_`, `data_loss_` and `scope_` )? Will that "eventual variable" be the new `Status` class ("new" as in after this PR) or the new `IOStatus` class or something else?

No. Only error_handler.cc needs to look at those additional fields, and it can cast the Status to IOStatus to access them.